maj hotline
Former-commit-id: 26c93f39ed9e48896036d708e832bf6d6e04fb3f

+++ htdocs/bimpcore/classes/BimpObject.php
@@ -771,9 +771,9 @@ class BimpObject extends BimpCache
     public function isLoaded()
     {
         if ($this->isDolObject()) {
-            return (isset($this->id) && $this->id && isset($this->dol_object->id) && $this->dol_object->id);
+            return (int) (isset($this->id) && $this->id && isset($this->dol_object->id) && $this->dol_object->id);
         }
-        return (isset($this->id) && $this->id);
+        return (int) (isset($this->id) && $this->id);
     }
 
     public function isNotLoaded()
